+++ app/components/clinic/PatientForm.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState, useMemo } from 'react';
 import PropTypes from 'prop-types';
-import { translate, Trans } from 'react-i18next';
+import { translate } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import compact from 'lodash/compact';
 import find from 'lodash/find';
@@ -19,7 +19,6 @@ import CloseRoundedIcon from '@material-ui/icons/CloseRounded';
 import CheckCircleRoundedIcon from '@material-ui/icons/CheckCircleRounded';
 import ErrorOutlineRoundedIcon from '@material-ui/icons/ErrorOutlineRounded';
 import { Box, Flex, Text, BoxProps } from 'rebass/styled-components';
-import sundial from 'sundial';
 import moment from 'moment';
 
 import * as actions from '../../redux/actions';
@@ -32,17 +31,10 @@ import { getCommonFormikFieldProps } from '../../core/forms';
 import { useIsFirstRender } from '../../core/hooks';
 import { dateRegex, patientSchema as validationSchema } from '../../core/clinicUtils';
 import { accountInfoFromClinicPatient } from '../../core/personutils';
-import { Body0, Body1, MediumTitle } from '../../components/elements/FontStyles';
+import { Body0 } from '../../components/elements/FontStyles';
 import { borders, colors } from '../../themes/baseTheme';
 import Icon from '../elements/Icon';
 import DexcomLogoIcon from '../../core/icons/DexcomLogo.svg';
-
-import {
-  Dialog,
-  DialogTitle,
-  DialogContent,
-  DialogActions,
-} from '../../components/elements/Dialog';
 
 function getFormValues(source, clinicPatientTags, disableDexcom) {
   const hasDexcomDataSource = !!find(source?.dataSources, { providerName: 'dexcom' });
@@ -75,7 +67,6 @@ export const PatientForm = (props) => {
   const { set: setToast } = useToasts();
   const selectedClinicId = useSelector((state) => state.blip.selectedClinicId);
   const clinic = useSelector(state => state.blip.clinics?.[selectedClinicId]);
-  const timePrefs = useSelector((state) => state.blip.timePrefs);
   const dateInputFormat = 'MM/DD/YYYY';
   const dateMaskFormat = dateInputFormat.replace(/[A-Z]/g, '9');
   const [initialValues, setInitialValues] = useState({});
@@ -88,15 +79,6 @@ export const PatientForm = (props) => {
   const showDexcomConnectState = !!selectedClinicId && !!dexcomDataSource?.state;
   const [showResendDexcomConnectRequest, setShowResendDexcomConnectRequest] = useState(false);
   const { sendingPatientDexcomConnectRequest } = useSelector((state) => state.blip.working);
-
-  const formattedLastRequestedDexcomConnectDate =
-    patient?.lastRequestedDexcomConnectTime &&
-    sundial.formatInTimezone(
-      patient?.lastRequestedDexcomConnectTime,
-      timePrefs?.timezoneName ||
-        new Intl.DateTimeFormat().resolvedOptions().timeZone,
-      'MM/DD/YYYY [at] h:mm a'
-    );
 
   const dexcomConnectStateUI = {
     pending: {
@@ -255,12 +237,12 @@ export const PatientForm = (props) => {
   }, [sendingPatientDexcomConnectRequest]);
 
   function handleResendDexcomConnectEmail() {
-    trackMetric('Clinic - Resend Dexcom connect email', { clinicId: selectedClinicId, dexcomConnectState })
+    trackMetric('Clinic - Resend Dexcom connect email', { clinicId: selectedClinicId, dexcomConnectState, source: 'patientForm' })
     setShowResendDexcomConnectRequest(true);
   }
 
   function handleResendDexcomConnectEmailConfirm() {
-    trackMetric('Clinic - Resend Dexcom connect email confirm', { clinicId: selectedClinicId });
+    trackMetric('Clinic - Resend Dexcom connect email confirm', { clinicId: selectedClinicId, source: 'patientForm' });
     formikContext.setStatus('resendingDexcomConnectRequest');
     dispatch(actions.async.sendPatientDexcomConnectRequest(api, selectedClinicId, patient.id));
   }
@@ -525,45 +507,15 @@ export const PatientForm = (props) => {
             </Body0>
           )}
 
-          <Dialog
-            id="resendDexcomConnectRequest"
-            aria-labelledby="dialog-title"
-            open={showResendDexcomConnectRequest}
+          <ResendDexcomConnectRequestDialog
+            api={api}
             onClose={() => setShowResendDexcomConnectRequest(false)}
-          >
-            <DialogTitle onClose={() => setShowResendDexcomConnectRequest(false)}>
-              <MediumTitle id="dialog-title">{t('Confirm Resending Connection Request')}</MediumTitle>
-            </DialogTitle>
-            <DialogContent>
-              <Body1>
-                {formattedLastRequestedDexcomConnectDate && (
-                  <Trans>
-                    <Text>
-                      You requested <Text as='span' fontWeight='bold'>{{patient: patient?.fullName || patient?.email}}</Text> to connect to <Text as='span' fontWeight='bold'>Dexcom</Text> on <Text as='span' fontWeight='bold'>{{requestDate: formattedLastRequestedDexcomConnectDate}}</Text>.
-                    </Text>
-                  </Trans>
-                )}
-                <Text>
-                  {t('Are you sure you want to resend this connection request?')}
-                </Text>
-              </Body1>
-            </DialogContent>
-            <DialogActions>
-              <Button variant="secondary" onClick={() => setShowResendDexcomConnectRequest(false)}>
-                {t('Cancel')}
-              </Button>
-              <Button
-                className="resend-dexcom-connect-request"
-                variant="primary"
-                processing={sendingPatientDexcomConnectRequest.inProgress}
-                onClick={() => {
-                  handleResendDexcomConnectEmailConfirm();
-                }}
-              >
-                {t('Resend Request')}
-              </Button>
-            </DialogActions>
-          </Dialog>
+            onConfirm={handleResendDexcomConnectEmailConfirm}
+            open={showResendDexcomConnectRequest}
+            patient={patient}
+            t={t}
+            trackMetric={trackMetric}
+          />
         </Box>
       )}
